requirements.txt
openai-whisper
torch
numpy
sounddevice
scipy
openai
chromadb
elevenlabs
vosk
pyttsx3
sentence-transformers
faiss-cpu
transformers
numpy